## Are you removing, renaming or moving a folder in this PR?
- [x] No, I'm not changing any folders (skip to TeamSites and delete the rest of this section)
- [ ] Yes, I'm removing, renaming or moving a folder

Did you change site-wide styles, platform utilities or other infrastructure?
- [x] No
- [ ] Yes, and I used the [proxy-rewrite steps](https://github.com/department-of-veterans-affairs/vets-website/tree/main/src/applications/proxy-rewrite#that-sounds-normal-so-whats-the-proxy-all-about) to test the injected header scenario

## Summary

This PR performs some basic updates to the `ApplicationDownloadLink` component in the Caregivers application for performance. Much of the work focuses around limiting recalucation of variables on re-render by utilizing `useCallback` and `useMemo` where necessary. The PR also updates the PDF fetch to async/await instead of Promise chaining. The was completed for better readability.

Based on testing done utilizing the React Profiler API the following gains were made:
- Base render time improved from 1.5ms to 0.9ms
- Re-render time improved from 0.8ms to 0.09ms 

## Related issue(s)

department-of-veterans-affairs/va.gov-team#103059

## Acceptance criteria

- Component is as performant as possible in all environments.

### Quality Assurance & Testing

- [x] No sensitive information (i.e. PII/credentials/internal URLs/etc.) is captured in logging, hardcoded, or specs
- [x] Linting warnings have been addressed

### Error Handling

- [x] Browser console contains no warnings or errors.
- [x] Events are being sent to the appropriate logging solution